+++ actix-rt/CHANGES.md
@@ -6,6 +6,18 @@
 
 * Add `System::attach_to_tokio` method. [#173]
 
+### Changed
+
+* Remove `'static` lifetime requirement for `Runtime::block_on` and `SystemRunner::block_on`. 
+  Remove `'static` lifetime requirement for `System::run` and `Builder::run`.
+  `Arbiter::spawn` would panic when `System` is not in scope. [#207]
+
+### Fixed
+
+* Fix work load issue by removing `PENDDING` thread local. [#207]
+
+[#207]: https://github.com/actix/actix-net/pull/207
+
 ## [1.1.1] - 2020-04-30
 
 ### Fixed